fix: Add graceful handling for component version mismatch and invalid IDs (#289)
* fix: Always return a valid app object

The `app` array must be rendered even if the client-claimed version is
newer than the server one

* fix: Update server tests

* fix: Rename FilterForUpdates to ProcessExtensionRequests

The old function name and its description were confusing - the goal is
to return a list of actions for **ALL** requested extensions, not only
those that have updates

* fix: Update v4 response to include information about all requested app IDs

Multiple extensions can be requested in the scope of a single request.
Until now, the unknown and blacklisted ones were not included in the
response.

* fix: SHA256 checks make sense only when there's a pending update

* fix: Incorporate v4 status changes into v3 impl

* fix: Reflect new logic in v3 response tests

* feat: Update v3 (XML/JSON) and v4 (JSON) UpdateResponse tests

* fix: Update v4 response tests

The empty extension list case was totally not realistic

* fix: Align v3 response tests with v4 ones